requirements.txt
openai==0.27.2
pydub==0.25.1
python-telegram-bot==20.2